requirements.txt
tensorflow>=2.5
google-cloud-vision>=2.0.0
numpy>=1.19
Pillow>=8.0
opencv-python>=4.5
matplotlib>=3.3